# fsmonitor updates for improved performance