This change does two main things.
1. Adds a check for the `CE_FSMONITOR_VALID` flag in the `ce_uptodate` macro so that whenever the code is checking if any entry is up to date the fsmonitor flag will be taken into consideration.
2. When unpacking trees keep the fsmonitor data so that the next command does not have to pay the price to check all the entries.