@kewillford 2019-10-31
Here are some of the performance differences with these changes

command|Previous Duration|Current Duration|Seconds difference|percent change
-|-|-|-|-|
checkout after folderplaceholder enumeration|17.40225|3.21683|-14.18542|-81.51486
status 1 after checkout feature/gvfs/perftest/defaultBranch|15.53006|3.03953|-12.49053|-80.42809
status 1 after checkout -b user/me/topic feature/gvfs/perftest/defaultBranch~1|15.66574|3.09206|-12.57368|-80.26228
status 1 after pull --ff-only origin feature/gvfs/perftest/checkoutBranch|16.08161|3.62207|-12.45954|-77.47694
status 1 after checkout -b pullTest feature/gvfs/perftest/checkoutBranch~10|15.61599|3.63645|-11.97954|-76.71329
status 1 after merge feature/gvfs/perftest/defaultBranch --no-commit|13.1538|3.17194|-9.98186|-75.88575
git reset --hard HEAD|26.76172|6.68994|-20.07178|-75.00183
status 1 after reset --hard HEAD|12.11019|3.15032|-8.95987|-73.98621
status 1 after checkout feature/gvfs/perftest/checkoutBranch|17.4508|4.88049|-12.57031|-72.03286
status 1 after stash|10.89549|3.09768|-7.79781|-71.56915
git add --all|30.97013|10.36075|-20.60938|-66.54599
status 2 after First ReadFiles|4.35711|1.60531|-2.7518|-63.15654
git stash|31.56066|11.86821|-19.69245|-62.39556
git merge feature/gvfs/perftest/defaultBranch --no-commit|21.15363|8.60551|-12.54812|-59.31899
git stash pop|19.23136|9.56369|-9.66767|-50.27034
git rebase feature/gvfs/perftest/defaultBranch|26.31549|13.19811|-13.11738|-49.84661
status 1 after merge --abort|14.68546|7.55865|-7.12681|-48.5297