@jeffhostetler 2019-11-01
@derrickstolee    Can you run this with GIT_TRACE_FSMONITOR turned on too?  Not sure if that'll help or not, but worth a shot.