@derrickstolee 2019-11-06
@kewillford when you get this rebased on top of `features/sparse-checkout-2.24.0` then I can launch the C# tests using watchman for additional confidence.