@derrickstolee 2019-11-12
@kewillford I kicked off a new build for you [here](https://dev.azure.com/gvfs/ci/_build/results?buildId=17538&view=results).